scheduler/example_scheduler_depth_app/requirements.txt
corva-sdk==1.8.0
#pytest==7.1.1
